+++ Takeout.java
@@ -1,13 +1,22 @@
 import java.util.StringTokenizer;
 import java.util.ArrayList;
+//import java.text.NumberFormat;
 
 //Create classes or function for all the random print statements
 
 class Takeout{
     public static void main(String[] args)
     {
+        //Menu Size
+        int menuSize = 5;
+
         //Example menu
-        String menu = "Chicken, 9.50\n" + "Steak, 10.50\n" + "Fish, 8.50\n" + "Vegetarian, 6.00\n" + "Fries, 3.00\n";
+        CustomArrayList menu = new CustomArrayList(menuSize);
+        menu.addMenuItem("Chicken", 9.50);
+        menu.addMenuItem("Steak", 10.50);
+        menu.addMenuItem("Fish", 8.50);
+        menu.addMenuItem("Vegetarian", 6.00);
+        menu.addMenuItem("Fries", 3.00);
 
         //Global variable initializations
         String readyStatus, order;
@@ -25,13 +34,12 @@ class Takeout{
         //Print Greeting for Customer
         System.out.println("\nHello, " + customerName +
         ", Welcome once again to Jeff's Restaurant! \n" +
-        "Please take a minute to look over our menu options.\n\n" + 
+        "Please take a minute to look over our menu options.\n"); 
 
         //Print out Menu
-        menu +
-
+        menu.printMenu(menu);
         //Prompt user to select if they are ready
-        "\nWhen you are ready, just type in \"ready\"\n" + 
+        System.out.println("\nWhen you are ready, just type in \"ready\"\n" + 
         "If you would like to see menu again type in \"menu\"\n" + 
         "or if you would like to leave the restaurant type in \"exit\"");
 
@@ -41,7 +49,7 @@ class Takeout{
             readyStatus = System.console().readLine().toLowerCase();
 
             if(readyStatus.equals("ready")){
-                System.out.print("\n\n" + customerName + " is ready to order.\nPlease place your order now with spaces in between items like so.\nex. Fries Steak\n\nOrder: ");
+                System.out.print("\n" + customerName + " is ready to order.\nPlease place your order now with spaces in between items like so.\nex. Fries Steak\n\nOrder: ");
                 
                 //Prompt Customer to input order
                 order = System.console().readLine();
@@ -58,12 +66,12 @@ class Takeout{
                 //******************************************ADD A LOOP TO CHECK IF IT IS CORRECT */
 
                 // Calculate Cost of Customer Order
-                orderCost = calcCost(fullOrder);
+                orderCost = calcCost(fullOrder,menu);
 
                 // Tell the Customer the cost of the order
                 System.out.println("Your order will cost: " + orderCost);
 
-                // Prompt Customer for amount of payment, and check to make sure it is sufficient funds
+                //************* */ Prompt Customer for amount of payment, and check to make sure it is sufficient funds
                 System.out.print("Please give amount that you are going to pay: ");
                 customerPayment = Double.parseDouble(System.console().readLine());
                 System.out.println(customerPayment);
@@ -80,7 +88,6 @@ class Takeout{
 
                 System.exit(1);
 
-
             } else if (readyStatus.equals("menu")) {
                 // Reprint Menu
                 System.out.println(menu + "\n");
@@ -95,7 +102,6 @@ class Takeout{
                     + "or if you would like to leave the restaurant type in \"exit\"");
         }
     }
-
 //Function to store order into an ArrayList
     static ArrayList<String> storeOrder(String order) {
         StringTokenizer orderToken = new StringTokenizer(order, " ");
@@ -108,11 +114,16 @@ class Takeout{
     }
 
 //Function to calculate cost of Food Order
-    static double calcCost(ArrayList<String> order){
+    static double calcCost(ArrayList<String> order,CustomArrayList menu){
         double cost = 0;
         
-        //Find a way to calculate cost of order
+        for(int i = 0; i<order.size(); i++){
+            for(int j = 0; j<menu.size(); j++){
+                if(order.get(i).equals(menu.get(j).name)){
+                    cost = cost + menu.get(j).cost;
+                }
+            } 
+        }
         return cost;
-
     }
 }