+++ Changelog.md
@@ -1,4 +1,16 @@
 # Version 0.1.1
+## Change
+* Create custom food class
+* Store Menu from Restaurant in a Custom Array List
+* Polish CustomArrayList class to have the following:
+    * addMenuItem(name,cost)
+    * get(index i)
+    * size()
+    * printMenu()
+* Create Storage Function to parse user input
+* Create Cost of order Function
+
+# Version 0.1.1
 ## Added
 * Rollback Takeout.java via new branch to a more simplistic form to break project into pieces
 * Changelog.md